+++ src/index.js
@@ -1,34 +1,58 @@
-import {classic, declarative} from './components/declarative';
+import {classic} from './components/declarative'
 
-const components = { Form: classic.Form };
-
-export {components, classic, declarative}
-
-/** deprecated */
-// Базовые копоненты
-export {default as CommandPanel} from './components/deprecated/CommandPanel/CommandPanel';
-export {default as FilterPanel} from './components/deprecated/FilterPanel/FilterPanel';
-export {default as Table} from './components/deprecated/Table/Table';
-export {default as AdvancedTable} from './components/deprecated/AdvancedTable/AdvancedTable';
-export {default as List} from './components/deprecated/List/List';
-export {default as Select} from './components/deprecated/Select/Select';
-export {default as SingleDate} from './components/deprecated/SingleDate/SingleDate';
-export {default as DateRange} from './components/deprecated/DateRange/DateRange';
-
-// Компоненты формы
-export {default as Form} from './components/deprecated/Form/Form';
-export {default as FileManager} from './components/deprecated/FileManager/FileManager';
-export {default as Modal} from './components/deprecated/Modal/Modal';
-
+export {deprecated} from './components/deprecated'
 // Store
-export { default as rtdReducer} from './redux/rtd.reducer';
+export {default as rtdReducer} from './redux/rtd.reducer';
 export {setDateStore} from './redux/rtd.actions';
-
 // Utils
 export {notificationError} from './components/utils/baseUtils'
-export { executeRequest } from "./components/utils/api";
-export * from './components/utils/datesUtils';
+export {executeRequest} from "./components/utils/api";
+// export * from './components/utils/datesUtils';
 
-
-
-
+// General
+export const { Button } = classic
+export const { Title } = classic
+export const { Text } = classic
+export const { DateText } = classic
+// Layout
+export const { Divider } = classic
+export const { Layout } = classic
+export const { Row } = classic
+export const { Col } = classic
+export const { Space } = classic
+// Data Entry
+export const { Checkbox } = classic
+export const { DatePicker } = classic
+export const { Form } = classic
+export const { FormHeader } = classic
+export const { FormBody } = classic
+export const { FormFooter } = classic
+export const { FormList } = classic
+export const { InputNumber } = classic
+export const { Input } = classic
+export const { Search } = classic
+export const { TextArea } = classic
+export const { Password } = classic
+export const { RadioGroup } = classic
+export const { Switch } = classic
+export const { Slider } = classic
+export const { Select } = classic
+export const { TreeSelect } = classic
+export const { Transfer } = classic
+export const { TimePicker } = classic
+export const { UploadFile } = classic
+// Data Display
+export const { Collapse } = classic
+export const { List } = classic
+export const { Popover } = classic
+export const { Tooltip } = classic
+export const { Tabs } = classic
+export const { TabPane } = classic
+// Feedback
+export const { Modal } = classic
+// Rt-design
+export const { Table } = classic
+export const { RtTable } = classic
+export const { AntTable } = classic
+export const { Custom } = classic
+export const { Switcher } = classic